Extract thin logic from runner into new ThinRunner
